pyproject.toml
[tool.poetry]
name = "goit-pythonweb-hw-12"
version = "0.1.0"
description = ""
authors = ["Ivan Dovhal <limetrack@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.6"
sqlalchemy = "^2.0.37"
alembic = "^1.14.0"
pydantic = {extras = ["email"], version = "^2.10.5"}
uvicorn = "^0.34.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
libgravatar = "^1.0.4"
python-multipart = "^0.0.20"
asyncpg = "^0.30.0"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.7.1"
slowapi = "^0.1.9"
fastapi-mail = "^1.4.2"
cloudinary = "^1.42.1"
redis = "^5.2.1"
pytest = "^8.3.4"
pytest-asyncio = "^0.25.3"
aiosqlite = "^0.20.0"


[tool.poetry.group.dev.dependencies]
pytest-cov = "^6.0.0"
httpx = "^0.28.1"
sphinx = "^8.1.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = "."
filterwarnings = "ignore::DeprecationWarning"
asyncio_default_fixture_loop_scope = "function"
